[SCF-70] - useCustomQuery 사용컴포넌트에 직접작성
